requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-PyMongo==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
PyJWT==2.8.0
pymongo==4.6.3
python-dotenv==1.0.1
Werkzeug==3.0.2
